fix: :truck: Moved gameList to components
Issue: #130
