+++ src/components/gameList/index.module.css
@@ -1,0 +1,68 @@
+.games {
+  padding: 1rem;
+}
+
+.games .list {
+  display: grid;
+	grid-template-columns: repeat(auto-fill, minmax(160px, 1fr));
+  gap: 1rem;
+  padding: 1rem 0;
+}
+
+/* .games .list .item {} */
+
+.games .list .item .thumbnail {
+  position: relative;
+  padding-bottom: 80%;
+}
+
+.games .list .item .thumbnail img {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+}
+
+.games .list .item .title {
+  font-size: 0.8rem;
+  font-weight: bold;
+  line-height: 1.6;
+}
+
+.games .list .item .title a {
+  color: black;
+  text-decoration: none;
+}
+
+.games .list .item .title a:hover {
+  cursor: pointer;
+  text-decoration: underline;
+}
+
+.games .list .item .description {
+  font-size: 0.7em;
+  line-height: 1.4;
+}
+
+.games .list .item .size {
+  font-size: 0.7rem;
+  line-height: 1.5;
+  color: var(--color-text-secondary);
+}
+
+.games .list .item .date {
+  color: var(--color-text-secondary);
+}
+
+.games .list .item .date p {
+  font-size: 0.7rem;
+  line-height: 1.5;
+  white-space: nowrap;
+}
+
+.games .list .item .date svg {
+  margin-right: 0.25rem;
+}
+
